task-87972-Commong action button functionality

+++ dashboard/views/_partial/action-buttons.php
@@ -7,13 +7,13 @@ if (isset($htmlContent) && $htmlContent != '') {
 }
 $btnGrp = $div->appendElement('div', array("class" => "btn-group"));
 $msg = isset($msg) ? $msg : '';
-if ((!isset($statusButtons) || true === $statusButtons)) {
+if (isset($statusButtons) && true === $statusButtons) {
     $div->appendElement('a', array('href' => 'javascript:void(0)', 'class' => 'btn btn-outline-brand btn-sm formActionBtn-js disabled', 'title' => Labels::getLabel('LBL_Publish', $siteLangId), "onclick" => "toggleBulkStatues(1, '" . $msg . "')"), '<i class="fas fa-eye"></i>', true);
 
     $div->appendElement('a', array('href' => 'javascript:void(0)', 'class' => 'btn btn-outline-brand btn-sm formActionBtn-js disabled', 'title' => Labels::getLabel('LBL_Unpublish', $siteLangId), "onclick" => "toggleBulkStatues(0, '" . $msg . "')"), '<i class="fas fa-eye-slash"></i>', true);
 }
 
-if (!isset($deleteButton) || true === $deleteButton) {
+if (isset($deleteButton) && true === $deleteButton) {
     $div->appendElement('a', array('href' => 'javascript:void(0)', 'class' => 'btn btn-outline-brand btn-sm formActionBtn-js disabled', 'title' => Labels::getLabel('LBL_Delete', $siteLangId), "onclick" => "deleteSelected()"), '<i class="fas fa-trash"></i>', true);
 }
